Fix: fix errors from renaming
