+++ .gitignore
@@ -14,6 +14,8 @@
 # torch related
 **/lightning_logs
 **/*.ckpt
+**/checkpoints
+**/csv_logs
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
