+++ tests/functional/dataset/patching_strategies/test_stratified_patching.py
@@ -3,7 +3,7 @@ from collections.abc import Sequence
 
 import numpy as np
 import pytest
-from tests.functional.dataset_ng.utils import (
+from tests.functional.dataset.utils import (
     assert_expected_pixel_probability,
     track_patching,
 )
