+++ tests/functional/dataset/test_filter_bg.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 
 import numpy as np
 import pytest
-from tests.functional.dataset_ng.utils import (
+from tests.functional.dataset.utils import (
     assert_expected_pixel_probability,
 )
